requirements.txt
asgiref==3.7.2
Django==5.0.2
django-admin==2.0.2
django-environ==0.11.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
excel-base==1.0.4
greenlet==3.0.3
isoweek==1.3.3
msgpack==1.0.7
neovim==0.3.1
psycopg2==2.9.9
PyJWT==2.8.0
pynvim==0.5.0
python-dateutil==2.8.2
pytz==2024.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tzlocal==5.2
xlwt==1.3.0
